refactor gulp tasks and utils
