+++ config/gulp/utils/env-vars.js
@@ -1,7 +1,7 @@
 var ENV_VARS;
 
 try {
-    ENV_VARS = require('../../env.json');
+    ENV_VARS = require('../../../env.json');
 } catch(e) {}
 
 module.exports = ENV_VARS;
